backend-api/requirements.txt
# KeyCare-Gemini3 Backend Dependencies
# =====================================

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Environment & Config
python-dotenv>=1.0.0

# HTTP Client (for Gemini API calls)
httpx>=0.26.0

# Google Gemini SDK (official)
google-generativeai>=0.4.0

# Validation
pydantic>=2.5.0

# CORS
# (included in fastapi)

# Optional: Rate limiting
slowapi>=0.1.9
